+++ packages/amplify-python-function-runtime-provider/src/util/packageUtils.ts
@@ -1,8 +1,9 @@
+import { execWithOutputAsString } from '@aws-amplify/amplify-cli-core';
 import { PackageRequest, PackageResult, ZipEntry } from '@aws-amplify/amplify-function-plugin-interface';
 import * as fs from 'fs-extra';
 import glob from 'glob';
 import * as path from 'path';
-import { execAsStringPromise, getPipenvDir, getPythonBinaryName, majMinPyVersion } from './pyUtils';
+import { getPipenvDir, getPythonBinaryName, majMinPyVersion } from './pyUtils';
 
 // packages python lambda functions and writes the archive to the specified file
 export async function pythonPackage(context: any, params: PackageRequest): Promise<PackageResult> {
@@ -11,7 +12,7 @@ export async function pythonPackage(context: any, params: PackageRequest): Promi
     const zipEntries: ZipEntry[] = [];
     if (params.service) {
       const pyBinary = getPythonBinaryName();
-      const pyVersion = await execAsStringPromise(`${pyBinary} --version`);
+      const pyVersion = await execWithOutputAsString(`${pyBinary} --version`);
       const layerPythonPath = path.join(params.srcRoot, 'lib', 'python' + majMinPyVersion(pyVersion), 'site-packages');
       const pipEnvDir = await getPipenvDir(params.srcRoot);
       // copy from virtualenv to layer path to maintain layer required structure