+++ packages/amplify-python-function-runtime-provider/src/util/pyUtils.ts
@@ -1,14 +1,12 @@
 import path from 'path';
 import fs from 'fs-extra';
-import { ExecOptions } from 'child_process';
-import execa from 'execa';
 import * as which from 'which';
 import { parse } from 'ini';
-import { AmplifyError } from '@aws-amplify/amplify-cli-core';
+import { AmplifyError, execWithOutputAsString } from '@aws-amplify/amplify-cli-core';
 
 // Gets the pipenv dir where this function's dependencies are located
 export async function getPipenvDir(srcRoot: string): Promise<string> {
-  const pipEnvDir = await execAsStringPromise('pipenv --venv', { cwd: srcRoot });
+  const pipEnvDir = await execWithOutputAsString('pipenv --venv', { cwd: srcRoot });
   const pyBinary = getPythonBinaryName();
 
   if (!pyBinary) {
@@ -31,22 +29,6 @@ export function majMinPyVersion(pyVersion: string): string {
   }
   const versionNum = pyVersion.split(' ')[1];
   return versionNum.split('.').slice(0, 2).join('.');
-}
-
-// wrapper for executing a shell command and returning the result as a string promise
-// opts are passed directly to the exec command
-export async function execAsStringPromise(command: string, opts?: ExecOptions): Promise<string> {
-  try {
-    let stdout = (await execa.command(command, opts)).stdout;
-
-    if (stdout) {
-      stdout = stdout.trim();
-    }
-
-    return stdout;
-  } catch (err) {
-    throw new AmplifyError('PackagingLambdaFunctionError', { message: `Received error [${err}] running command [${command}]` });
-  }
 }
 
 export const getPythonBinaryName = (): string | undefined => {